tablenotfound, generic table auth protection
