+++ src/providers/AuthProvider.tsx
@@ -1,0 +1,40 @@
+import React, { createContext, useContext, useEffect, useState } from "react";
+import { onAuthStateChanged } from "firebase/auth";
+import { useFirebaseAuth } from "../lib/firebase/hooks/useFirebase";
+
+
+interface UserType {
+  email: string | null;
+  uid: string | null;
+}
+
+const AuthContext = createContext({});
+
+export const useUser = () => useContext<any>(AuthContext);
+
+export const AuthContextProvider = ({ children }: { children: React.ReactNode }) => {
+  const [user, setUser] = useState<UserType>({ email: null, uid: null });
+  const [loading, setLoading] = useState<boolean>(true);
+  const auth = useFirebaseAuth();
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChanged(auth, (user) => {
+      if (user) {
+        setUser({
+          email: user.email,
+          uid: user.uid,
+        });
+      } else {
+        setUser({ email: null, uid: null });
+      }
+    });
+    setLoading(false);
+    return unsubscribe;
+  }, []);
+
+  return (
+    <AuthContext.Provider value={{ user }}>
+      {children}
+    </AuthContext.Provider>
+  );
+}
